code/RAGAgentServer/requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit
langchain-community
langserve
uvicorn[standard]
google-generativeai
bs4
langchain-google-genai
cassio
langchainhub
duckduckgo-search
google-search-results
xmltodict
langchain_experimental
pydantic
crewai
crewai_tools
Flask
tavily-python
python_utils
Flask-Cors
groq
langchain-groq
chromadb
html2text
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
tiktoken
faiss-cpu
playwright
pandas
fastapi
gunicorn
Werkzeug
pymongo

